removed blacklisted token

+++ MisuiIndia/authentication/models.py
@@ -3,7 +3,7 @@ from django.db import models
 from django.contrib.auth.models import AbstractUser
 
 class CustomUser(AbstractUser):
-    contact_number = models.IntegerField(null=True, blank = True) #Edit this null part
+    contact_number = models.BigIntegerField(null=True, blank = True) #Edit this null part
     isSeller = models.BooleanField(default = False )
     adharCardNumber = models.BigIntegerField(unique = True, null = True, blank = True)
     PanNumber = models.CharField(unique  = True,  null = True, blank = True, max_length = 10)
